Fix native build issue perception_3nodes, prototype offloaded
These changes allow for native builds of the whole acceleration_examples
package fixing a previous issue with perception_3nodes package. In
addition, it prototypes the build of the accelerated binaries from
offloaded_doublevadd_publisher. This could be used for native on-target
builds (e.g. in Ubuntu 22.04) and answers some user demands that have
appeared

Signed-off-by: Víctor Mayoral Vilches <v.mayoralv@gmail.com>